all_requirements.txt
absl-py==0.1.12
alabaster==0.7.10
asn1crypto==0.24.0
astor==0.6.2
astroid==1.6.2
Babel==2.5.3
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==1.5.0
bokeh==0.12.15
boto==2.48.0
boto3==1.6.16
botocore==1.9.16
bs4==0.0.1
bz2file==0.98
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
colorama==0.3.9
console==0.1.1
cryptography==2.2.1
cycler==0.10.0
Cython==0.28.1
decorator==4.2.1
docutils==0.14
entrypoints==0.2.3
Flask==1.0.2
future==0.16.0
gast==0.2.0
gensim==3.4.0
grpcio==1.10.0
gunicorn==19.8.1
html5lib==0.9999999
idna==2.6
imagesize==1.0.0
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.0
isort==4.3.4
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
line-bot-sdk==1.5.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
notebook==5.4.1
numpy==1.14.2
numpydoc==0.7.0
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pickleshare==0.7.4
pip-review==1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psutil==5.4.3
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.8
python-crfsuite==0.9.5
python-dateutil==2.7.2
pytz==2018.3
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.0
requests==2.18.4
rope==0.10.7
s3transfer==0.1.13
Sastrawi==1.0.1
scikit-learn==0.19.1
scipy==1.0.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
sklearn-crfsuite==0.3.6
smart-open==1.5.7
snowballstemmer==1.2.1
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
tabulate==0.8.2
tb-nightly==1.8.0a20180415
tensorboard==1.6.0
tensorflow-gpu==1.6.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tf-nightly-gpu==1.8.0.dev20180329
tornado==5.0.2
tqdm==4.23.2
traitlets==4.3.2
typing==3.6.4
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.0
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
